requirements.txt
pandas~=1.4.2
requests~=2.28.0
beautifulsoup4~=4.11.1
python-dotenv~=0.20.0
SQLAlchemy~=1.4.37